refactor: rollback module restructure
