+++ src/agent.rs
@@ -63,11 +63,11 @@ Note that there is an Oracle and Exporter for each network, but only one Local S
 ################################################################################################################################## */
 
 pub mod dashboard;
+pub mod market_hours;
 pub mod market_schedule;
 pub mod metrics;
 pub mod pythd;
 pub mod remote_keypair_loader;
-pub mod schedule;
 pub mod solana;
 pub mod store;
 use {
